% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddsDownload.R
\name{ddsDownload}
\alias{ddsDownload}
\title{DDS RESTful File Download.}
\usage{
ddsDownload(project = NULL, local_folder = NULL, stop_on_mismatch = TRUE,
  chunk_size_bytes = 1293892)
}
\arguments{
\item{project}{The DDS project to download.}

\item{local_folder}{The location (UNC path) of a local folder where files will be saved}

\item{chunk_size_bytes}{optionally, files can be downloaded per chunck, specify download chunk size here.}

\item{stop_on_mismatch=TRUE}{A logical indicating if ddsDownload should abort due to mistached DDS hash and current download}
}
\value{
Standard Output indicating a successful or failed download.
}
\description{
DDS RESTful File Download.
}
\examples{
ddsDownload(local_folder="/Users/nn31/Desktop/upload_testers",project="UploadTester")
}

