% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/login.R
\name{ddslogin}
\alias{ddslogin}
\title{Login function to DDS environment.}
\usage{
ddslogin(url = NA, rememberMe = TRUE)
}
\arguments{
\item{url}{The URL to a valid DDS portal (PROD,DEV,UATEST).}

\item{rememberME}{A logical indicating if the session login information should be stored later.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Login function to DDS environment.
}
\examples{
ddslogin()
ddslogin(url='https://dukeds-uatest.herokuapp.com')
}

