% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddsRequest.R
\name{ddsRequest}
\alias{ddsRequest}
\title{DDS RESTful API wrapper function.}
\usage{
ddsRequest(url = paste0(.getCache("url"), "/api/v1"), endpoint = NULL,
  body_list = NULL, customrequest = "GET",
  httpheader = .getCache("curlheader"), curlHandle, pass = FALSE)
}
\arguments{
\item{url}{The URL to a valid DDS portal (PROD,DEV,UATEST).}

\item{endpoint}{The DDS endpoint to call out to}

\item{body_list}{Takes an R list to supply as the body of a call}

\item{customrequest}{One of the following ('GET','PUT','POST')}

\item{httpheader}{Takes an R list to supply as a modified header of a call}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
DDS RESTful API wrapper function.
}
\examples{
ddsLogin()
ddsLogin(url='https://dukeds-uatest.herokuapp.com')
}

