% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddsUpload.R
\name{ddsUpload}
\alias{ddsUpload}
\title{DDS RESTful File Upload.}
\usage{
ddsUpload(file_folder = NULL, project = NULL, chunk_size_bytes = 1293892)
}
\arguments{
\item{file_folder}{The location (UNC path) of a file or folder to be uploaded}

\item{project}{The DDS project to upload the payload.}

\item{chunk_size_bytes}{If determined that files need to be uploaded, specify upload chunk size here.}
}
\value{
The URL of the resource available on DDS.
}
\description{
DDS RESTful File Upload.
}
\examples{
ddsUpload(file_folder="/Users/nn/Desktop/uplo.ad_tester",project="UploadTester")
}

