% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/login.R
\name{ddsLogin}
\alias{ddsLogin}
\title{Login function to DDS environment.}
\usage{
ddsLogin(url = NA, rememberMe = TRUE)
}
\arguments{
\item{url}{The URL to a valid DDS portal (https://dukeds.herokuapp.com,https://dukeds-dev.herokuapp.com,https://dukeds-uatest.herokuapp.com). Dev will be used if no url is given.}

\item{rememberMe}{A logical indicating if the session login information should be stored later.}
}
\value{
ddsLogin will return with a message detailing where credentials can be found (i.e. variable curlheader).
}
\description{
Login function to DDS environment.
}
\examples{
ddsLogin()
ddsLogin(url='https://dukeds-uatest.herokuapp.com')
}

