% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/login.R
\name{ddslogout}
\alias{ddslogout}
\title{Function to sign out of DDS. This amounts to removing the authorization jwt.}
\usage{
ddslogout()
}
\description{
Function to sign out of DDS. This amounts to removing the authorization jwt.
}
\examples{
ddslogout()
}

